feat: [prd-website-semantic-color-token-consolidation-005] - [Enforce semantic token usage with an automated static guard]
